+++ src/Models/SecureHeadersMiddlewareConfiguration.cs
@@ -78,6 +78,11 @@ public class SecureHeadersMiddlewareConfiguration
     public bool UseCrossOriginEmbedderPolicy { get; set; }
 
     /// <summary>
+    /// Indicates whether the response should use the Reporting-Endpoints header
+    /// </summary>
+    public bool UseReportingEndPoints { get; set; }
+
+    /// <summary>
     /// Indicates whether the response should use Clear-Site-Data
     /// </summary>
     public bool UseClearSiteData { get; set; }
@@ -132,6 +137,8 @@ public class SecureHeadersMiddlewareConfiguration
     public CrossOriginOpenerPolicy CrossOriginOpenerPolicy { get; set; }
 
     public CrossOriginEmbedderPolicy CrossOriginEmbedderPolicy { get; set; }
+
+    public ReportingEndpointsPolicy ReportingEndpointsPolicy { get; set; }
 
     /// <summary>
     /// The Clear-Site-Data path configuration to use